Make template repository URL and ref modifiable
This commit changes the repository URLs and default branch names for
both program and policy pack template repositories from constants to
variables, in order that they are customizable by linker flags.

For example, compiling the Pulumi CLI with the following linker flags:

- `-X [module path]/workspace.pulumiTemplateGitRepository=https://github.com/jen20/pulumi-templates.git`
- `-X [module path]/workspace.pulumiTemplateBranch=main`

Would instruct the `pulumi new` command to pull the list of available
templates from the `main` branch of the specified repository instead of
the upstream default. The defaults are left unchanged.