Fix `pulumi refresh` diffs
Refresh diffs should be diffs of *outputs* not *inputs*.  The code was previously attempting to do this - but was not correct because OpRefresh transmogrifies itself into OpUpdate, OpSame or OpDelete as it is processed, and so those resulting operations (which were what shows the diff) would render their input diff as normal.

Instead, decide whether to render refresh-based output diffs based on whether the action being performed is a `pulumi refresh` instead of using the Op kind.

Related to #6243, #5713, #5816 and possibly others.