[cli] prevent panic in pulumi python applications when apply is used incorrectly
Fixes: #5236

This bug *only* affected Python - as other language runtimes caught the
incorrect use of apply at runtime. Given the following Pulumi program:

```
from pulumi_aws import iam, s3
my_bucket = s3.Bucket(
    "simple-bucket",
    force_destroy=True,
    acl="private",
    versioning=s3.BucketVersioningArgs(enabled=True),
)
policy_document = iam.get_policy_document(
    statements=[
        iam.GetPolicyDocumentStatementArgs(
            sid="",
            effect="Allow",
            actions=["s3:GetObject"],
            resources=[my_bucket.arn.apply(lambda arn: f"{arn}/*")],
        ),
    ]
)
```

The apply is incorrect here as the Invoke should be run as part of the apply. This
application will produce a panic message similar to the following:

```
pulumi:pulumi:Stack (my-stack-dev):
    panic: fatal: A failure has occurred: Unrecognized structpb value kind in RPC[tf.Provider[aws].Invoke(aws:iam/getPolicyDocument:getPolicyDocument).args]: <nil>
    goroutine 114 [running]:
    github.com/pulumi/pulumi/sdk/v2/go/common/util/contract.failfast(...)
    	/home/runner/go/pkg/mod/github.com/pulumi/pulumi/sdk/v2@v2.9.1-0.20200825190708-910aa96016cd/go/common/util/contract/failfast.go:23
    github.com/pulumi/pulumi/sdk/v2/go/common/util/contract.Failf(0x6bc984f, 0x2f, 0xc0010daec8, 0x2, 0x2)
```

In order to help the user, we are able to detect if the error is due to an Invoke and point
them in the direction of the `apply` docs:

```
    Exception: invoke of aws:iam/getPolicyDocument:getPolicyDocument failed: invocation of aws:iam/getPolicyDocument:getPolicyDocument returned an error:
    Incorrect use of `apply`. Please refer to https://www.pulumi.com/docs/intro/concepts/inputs-outputs/#apply for further details on how to use `apply`
    error: an unhandled error occurred: Program exited with non-zero exit code: 1
```