[sdk/python] Ensure Output objects are not iterable
Although Output objects can never correct support iteration, Python will see the implementation of `__getitem__` and try to iterate the object, leading to an infinite loop.  To prevent this, we need to explicitly implement __iter__ and make it return a TypeError to prevent iteration (and offer a useful error message).

Fixes #5028.