Re-enable tests. (#7462)
A number of tests have likely not been running since tags were added to
each test file. This is because `go list` uses build flags when scanning
files, so any pacakges that did not include at least one file that did
not have any tags (read: about half of the integration tests) were not
picked up by the command used in the Makefile to detect tests.