Add ProviderMain function
This commit adds a new counterpart to `ComponentMain` which accepts
functional options for specifying callback functions. Currently it
supports `construct` (for components) and `call` (for methods), but is
extensible in a non-breaking fashion in future to support all other
provider methods as they become desirable to implement.

The original `ComponentMain` still exists, though it may be desirable to
deprecate it in future in favor of `ProviderMain`.