[codegen/model] Speed up union types.
The goal of these changes is to reduce the allocation volume incurred by
union types.

Union types currently allocate a large volume of data when they are
created in order to deduplicate element types and sort element types
into a canonical order. The deduplication is important as it allows for
a fast path during comparison based on the number of elements, but the
sorting--also used to simplify equality checks--generally costs more
than it is worth, as it requires stringifying each element type.

These changes replace the string-based deduplication process with either
a simple O(N^2) approach for small numbers of element types or a more
complex hashtable-based approach for larger numbers of element types.
The same approach is used for equality checks.

These changes speed up the code generation process by decreasing the
amount of time spent in the allocator and garbage collector.