Rehydrate provider resources in `Construct`
Previously, any provider resource passed to multi-lang components would
be instantiated as a `DependencyProviderResource` inside `Construct`,
which prevents the component from being able to easily access the
provider's state as an instance of of the provider
(e.g. `*aws.Provider`).

This change attempts to rehydrate the provider resource in the same way
that resource references are rehydrated, if it's been registered, s.t.
the specific provider resource type is instantiated with its state.
Otherwise falling back to returning `DependencyProviderResource`.