[engine] expand dependencies in step gen
These changes implement a higher-level fix for #7780 by expanding the
dependencies of a resource as part of step generation.

Dependency expansion is responsible for expanding dependencies on
component resources into dependencies on the component resource _and_
the component resource's descendents _at the time the dependent is
registered_.

For example, given a dependency on Comp1 in the following parent/child
graph:

              Comp1
         _______|__________
         |      |          |
      Cust1   Comp2      Comp3
             ___|___       |
             |     |       |
           Cust2  Cust3  Comp4
             |             |
           Cust4         Cust5

The expanded dependencies set will be [Comp1, Cust1, Comp2, Comp3,
Cust2, Cust3, Comp4, Cust5]. Notably, Cust4 is *not* a member of the
set because it is a child of a custom resource.

These changes are complementary to the changes in #7788. The primary
benefit of the early expansion approach is that the expanded
dependencies will be present in events and the statefile, so downstream
consumers do not need to replicate the expansion logic in order to
understand the totality of a resource's dependencies.