Revert "Allow Python dynamic provider resources to be constructed outside of `__main__` (#7755)" (#7889)
This reverts commit ebb0e6aaed83a2275bed0284ec96fbb4ebfc3d67.

The changes in #7755 introduced a regression tracked in #7795.  It is not yet clear how to retain the desired behaviour introduced in #7755 while avoiding this regression, so for now we will revert those changes, and re-open #7453 to track a deeper fix.  That fix may require making changes to upstream `dill` to properly support these serialization requirements.

Fixes #7795.