Avoid unnecessary ToOutput conversions
Handle cases where the destination is a ptr type whose element type is
the same as the value type (e.g. destType is *FooBar and valueType is
FooBar). This avoids calling the ToOutput method to convert the input to
an output in this case. It also avoids infinite loops when an output's
value is an input (which can happen with enums).